fix ambiguous overloads by calling math routines (C++17)
Using gcc-11/12/13 on illumos platforms.

Maybe C++23 will fix it, but for now it compiles using standard C++17,
and not all overloads are implemented in the STL.

+++ libosmscout-client-qt/src/osmscoutclientqt/PlaneMapRenderer.cpp
@@ -281,7 +281,7 @@ bool PlaneMapRenderer::RenderMap(QPainter& painter,
 double PlaneMapRenderer::computeScale(const osmscout::MercatorProjection &previousProjection,
                                       const osmscout::MercatorProjection &currentProjection)
 {
-  double currentDiagonal=sqrt(pow(currentProjection.GetWidth(),2) + pow(currentProjection.GetHeight(),2));
+  double currentDiagonal=sqrt(pow(double(currentProjection.GetWidth()),2) + pow(double(currentProjection.GetHeight()),2));
 
   osmscout::GeoCoord topLeft;
   currentProjection.PixelToGeo(0,0,topLeft);
